Remove error code types from the VM
Instead the VM will now throw errors codes (regular integers) upon
encountering errors.

+++ vm/src/error_codes.rs
@@ -2,31 +2,31 @@ use std::io;
 
 // Generic errors
 
-pub const STRING_INVALID_UTF8: u16 = 1;
+pub const STRING_INVALID_UTF8: i64 = 1;
 
 // IO errors, ranging from 100 to 200.
 
-pub const IO_GENERIC: u16 = 100;
-pub const IO_PERMISSION_DENIED: u16 = 101;
-pub const IO_CONNECTION_REFUSED: u16 = 102;
-pub const IO_CONNECTION_RESET: u16 = 103;
-pub const IO_CONNECTION_ABORTED: u16 = 104;
-pub const IO_NOT_CONNECTED: u16 = 105;
-pub const IO_ADDRESS_IN_USE: u16 = 106;
-pub const IO_ADDRESS_NOT_AVAILABLE: u16 = 107;
-pub const IO_BROKEN_PIPE: u16 = 108;
-pub const IO_ALREADY_EXISTS: u16 = 109;
-pub const IO_WOULD_BLOCK: u16 = 110;
-pub const IO_INVALID_INPUT: u16 = 111;
-pub const IO_INVALID_DATA: u16 = 112;
-pub const IO_TIMED_OUT: u16 = 113;
-pub const IO_WRITE_ZERO: u16 = 114;
-pub const IO_INTERRUPTED: u16 = 115;
-pub const IO_UNEXPECTED_EOF: u16 = 116;
-pub const IO_NOT_FOUND: u16 = 117;
+pub const IO_GENERIC: i64 = 100;
+pub const IO_PERMISSION_DENIED: i64 = 101;
+pub const IO_CONNECTION_REFUSED: i64 = 102;
+pub const IO_CONNECTION_RESET: i64 = 103;
+pub const IO_CONNECTION_ABORTED: i64 = 104;
+pub const IO_NOT_CONNECTED: i64 = 105;
+pub const IO_ADDRESS_IN_USE: i64 = 106;
+pub const IO_ADDRESS_NOT_AVAILABLE: i64 = 107;
+pub const IO_BROKEN_PIPE: i64 = 108;
+pub const IO_ALREADY_EXISTS: i64 = 109;
+pub const IO_WOULD_BLOCK: i64 = 110;
+pub const IO_INVALID_INPUT: i64 = 111;
+pub const IO_INVALID_DATA: i64 = 112;
+pub const IO_TIMED_OUT: i64 = 113;
+pub const IO_WRITE_ZERO: i64 = 114;
+pub const IO_INTERRUPTED: i64 = 115;
+pub const IO_UNEXPECTED_EOF: i64 = 116;
+pub const IO_NOT_FOUND: i64 = 117;
 
 /// Returns an error code for a Rust IO error.
-pub fn from_io_error(error: io::Error) -> u16 {
+pub fn from_io_error(error: io::Error) -> i64 {
     match error.kind() {
         io::ErrorKind::NotFound => IO_NOT_FOUND,
         io::ErrorKind::PermissionDenied => IO_PERMISSION_DENIED,
